Update types to handle calling ObjectID without new
We use interfaces here to allow us to define an object as both "newable"
and callable. We have one interface for the class constructor, this
defines how to call it (new & as a function) as well as the static
methods. We then have another interface for the instance, this defines
the members and methods.

We then have to export a const ObjectID so that typescript knows we are
dealing with values, rather than types.

ref https://discord.com/channels/508357248330760243/765583369345302570/879753593320796230